+++ adapter/src/components/Alerts.js
@@ -25,9 +25,12 @@ const Alerts = () => {
         )
 
     useEffect(() => {
-        if (alertManagerAlerts.length > 0 || alertStackAlerts.length > 0) {
-            setAlertStackAlerts(
-                mergeAlertStackAlerts(alertStackAlerts, alertManagerAlerts)
+        if (alertManagerAlerts.length > 0) {
+            setAlertStackAlerts((currentAlertStackAlerts) =>
+                mergeAlertStackAlerts(
+                    currentAlertStackAlerts,
+                    alertManagerAlerts
+                )
             )
         }
     }, [alertManagerAlerts])