+++ cli/src/lib/generateManifests.js
@@ -1,6 +1,28 @@
 const { reporter, chalk } = require('@dhis2/cli-helpers-engine')
 const fs = require('fs-extra')
 
+/**
+ * Gets the original `entrypoints` property in d2.config.js
+ * without applying defaults. Used to detect if there is actually
+ * supposed to be an app entrypoint for this... app. Temporary until
+ * the build process is redesigned to allow building plugins without
+ * apps (LIBS-479)
+ */
+const getOriginalEntrypoints = (paths) => {
+    try {
+        if (fs.existsSync(paths.config)) {
+            reporter.debug('Loading d2 config at', paths.config)
+            // NB: this import can be confounded by previous object mutations
+            const originalConfig = require(paths.config)
+            reporter.debug('loaded', originalConfig)
+            return originalConfig.entryPoints // may be undefined
+        }
+    } catch (e) {
+        reporter.error('Failed to load d2 config!')
+        reporter.error(e)
+        process.exit(1)
+    }
+}
 const parseCustomAuthorities = (authorities) => {
     if (!authorities) {
         return undefined
@@ -93,10 +115,19 @@ module.exports = (paths, config, publicUrl) => {
             },
         ],
         start_url: '.',
-        display: 'browser',
+        display: 'standalone',
         theme_color: '#ffffff',
         background_color: '#f4f6f8',
     }
+
+    const includesPlugin = Boolean(config.entryPoints.plugin)
+    // If there's a plugin, there might not be an app intended to be exposed,
+    // in which case omit the app launch path. Check the original d2.config
+    // without added defaults to see if an app is intended.
+    // If there's not a plugin, default to 'true'
+    const shouldIncludeAppLaunchPath = includesPlugin
+        ? Boolean(getOriginalEntrypoints(paths)?.app)
+        : true
 
     // Legacy manifest
     const manifestWebapp = {
@@ -108,7 +139,8 @@ module.exports = (paths, config, publicUrl) => {
         version: config.version,
         core_app: config.coreApp,
 
-        launch_path: 'index.html',
+        launch_path: shouldIncludeAppLaunchPath ? paths.launchPath : undefined,
+        plugin_launch_path: includesPlugin ? paths.pluginLaunchPath : undefined,
         default_locale: 'en',
         activities: {
             dhis: {
@@ -146,6 +178,10 @@ module.exports = (paths, config, publicUrl) => {
     // Write d2 config json
     const appConfig = { ...config }
     delete appConfig['entryPoints']
+    appConfig.entryPoints = {
+        app: shouldIncludeAppLaunchPath ? paths.launchPath : undefined,
+        plugin: includesPlugin ? paths.pluginLaunchPath : undefined,
+    }
     delete appConfig['pwa']
 
     fs.writeJsonSync(paths.shellPublicConfigJson, appConfig, { spaces: 2 })