+++ cli/src/lib/paths.js
@@ -7,20 +7,20 @@ const shellSource = path.dirname(
 )
 const shellAppDirname = 'src/D2App'
 
-const findYarnLock = (base) => {
-    if (base === '/') {
+const findYarnLock = (base, rootDirectory) => {
+    if (base === rootDirectory) {
         return null
     }
-
     const yarnLock = path.join(base, './yarn.lock')
     if (fs.existsSync(yarnLock)) {
         return yarnLock
     }
-    return findYarnLock(path.dirname(base))
+    return findYarnLock(path.dirname(base), rootDirectory)
 }
 
 module.exports = (cwd = process.cwd()) => {
     const base = path.resolve(cwd)
+    const rootDirectory = path.parse(cwd).root
     const paths = {
         babelConfig: path.join(__dirname, '../../config/babel.config.js'),
         configDefaultsApp: path.join(
@@ -44,7 +44,7 @@ module.exports = (cwd = process.cwd()) => {
 
         base,
         package: path.join(base, './package.json'),
-        yarnLock: findYarnLock(base),
+        yarnLock: findYarnLock(base, rootDirectory),
         dotenv: path.join(base, './.env'),
         config: path.join(base, './d2.config.js'),
         readme: path.join(base, './README.md'),
@@ -57,14 +57,24 @@ module.exports = (cwd = process.cwd()) => {
         d2: path.join(base, './.d2/'),
         appOutputFilename: 'App.js',
         shell: path.join(base, './.d2/shell'),
+        shellSrc: path.join(base, './.d2/shell/src'),
         shellAppEntrypoint: path.join(base, './.d2/shell/src/App.js'),
         shellAppDirname,
         shellApp: path.join(base, `./.d2/shell/${shellAppDirname}`),
+        shellPluginBundleEntrypoint: path.join(
+            base,
+            './.d2/shell/src/plugin.index.js'
+        ),
+        shellPluginEntrypoint: path.join(base, './.d2/shell/src/Plugin.js'),
         shellSrcServiceWorker: path.join(
             base,
             './.d2/shell/src/service-worker.js'
         ),
         shellPublic: path.join(base, './.d2/shell/public'),
+        shellPublicPluginHtml: path.join(
+            base,
+            './.d2/shell/public/plugin.html'
+        ),
         shellPublicServiceWorker: path.join(
             base,
             './.d2/shell/public/service-worker.js'
@@ -101,6 +111,9 @@ module.exports = (cwd = process.cwd()) => {
         ),
         buildLibBundleFile: '{name}-{version}.zip',
         buildLibBundleOutput: path.join(base),
+
+        launchPath: 'index.html',
+        pluginLaunchPath: 'plugin.html',
     }
 
     reporter.debug('PATHS', paths)