+++ pwa/src/offline-interface/offline-interface.js
@@ -1,7 +1,51 @@
 import EventEmitter from 'events'
 import { swMsgs } from '../lib/constants.js'
-import { register, unregister, checkForUpdates } from '../lib/registration.js'
+import {
+    register,
+    unregister,
+    checkForUpdates,
+    getRegistrationState,
+} from '../lib/registration.js'
 import { openSectionsDB, SECTIONS_STORE } from '../lib/sections-db.js'
+
+const PWA_ENABLED = process.env.REACT_APP_DHIS2_APP_PWA_ENABLED === 'true'
+
+/**
+ * This and the following 'test' functions test for PWA features and log errors
+ * if there's an issue so they can be reused in the Offline Interface methods.
+ *
+ * Known situations when navigator.serviceWorker is not available:
+ * 1. Private browsing in firefox
+ * 2. Insecure contexts (e.g. http that's not local host)
+ */
+function testSWAvailable() {
+    if ('serviceWorker' in navigator) {
+        return true
+    }
+
+    const msg =
+        (!window.isSecureContext
+            ? 'This window is not a secure context -- see https://developer.mozilla.org/en-US/docs/Web/Security/Secure_Contexts.'
+            : '`serviceWorker` is not available on `navigator`.') +
+        ' PWA features will not work.'
+    console.error(new Error(msg))
+    return false
+}
+
+function testPWAEnabled() {
+    if (PWA_ENABLED) {
+        return true
+    }
+
+    const msg =
+        'PWA is not enabled in `d2.config.js`. No service worker will be registered and offline interface features will not work.'
+    console.error(new Error(msg))
+    return false
+}
+
+function testPWAAndSW() {
+    return testSWAvailable() && testPWAEnabled()
+}
 
 /** Helper to simplify SW message sending */
 function swMessage(type, payload) {
@@ -21,7 +65,12 @@ function swMessage(type, payload) {
  */
 export class OfflineInterface {
     constructor() {
-        this.pwaEnabled = process.env.REACT_APP_DHIS2_APP_PWA_ENABLED === 'true'
+        // Helper property that consumers can check
+        this.pwaEnabled = PWA_ENABLED
+
+        // The latest value from the service worker. The `this.ready` promise
+        // will resolve when this gets a boolean value from the SW
+        this.latestIsConnected = null
 
         if (this.pwaEnabled) {
             register()
@@ -29,7 +78,11 @@ export class OfflineInterface {
             unregister()
         }
 
-        if (!('serviceWorker' in navigator)) {
+        if (!testSWAvailable()) {
+            // Make this value available for ServerVersionProvider
+            this.ready = Promise.resolve()
+            // Must be connected if we can't use a SW for offline features
+            this.latestIsConnected = true
             return
         }
 
@@ -56,6 +109,44 @@ export class OfflineInterface {
             this.offlineEvents.emit(type, payload)
         }
         navigator.serviceWorker.addEventListener('message', handleSWMessage)
+
+        // (todo: refactor to another function)
+        // When this promise resolves, it indicates that a connection status
+        // value has been received from the service worker and is available
+        // as a property on this offlineInterface.
+        // Expected to be used by ServerVersionProvider in the app adapter
+        // to delay rendering the app-runtime Provider until ready.
+        this.ready = !this.pwaEnabled
+            ? Promise.resolve()
+            : new Promise((resolve) => {
+                  // Listen to status updates and store the latest value here so the
+                  // connection status hook can initialize to this value
+                  this.offlineEvents.on(
+                      swMsgs.dhis2ConnectionStatusUpdate,
+                      ({ isConnected }) => {
+                          // If this is the first time receiving an update from the
+                          // SW, resolve the this.ready promise
+                          const shouldResolveReady =
+                              this.latestIsConnected === null
+                          this.latestIsConnected = isConnected
+                          if (shouldResolveReady) {
+                              resolve()
+                          }
+                      }
+                  )
+
+                  try {
+                      // Prompt the SW to send back connection status
+                      // without its usual delay
+                      swMessage(swMsgs.getImmediateDhis2ConnectionStatusUpdate)
+                  } catch {
+                      // It's likely the SW hasn't installed yet, so go ahead and
+                      // resolve `ready` -- the app must be online to get to this case
+                      // anyway
+                      this.latestIsConnected = true
+                      resolve()
+                  }
+              })
     }
 
     /** Basically `checkForUpdates` from registration.js exposed here */
@@ -64,18 +155,28 @@ export class OfflineInterface {
         checkForUpdates({ onUpdate: onNewSW })
     }
 
+    async getRegistrationState() {
+        return await getRegistrationState()
+    }
+
     /**
      * Requests clients info from the active service worker.
      * @returns {Promise}
      */
     getClientsInfo() {
+        if (!testSWAvailable()) {
+            return Promise.resolve({ clientsCount: 0 })
+        }
+
         return new Promise((resolve, reject) => {
             navigator.serviceWorker.getRegistration().then((registration) => {
-                if (!registration || !registration.active) {
-                    reject('There is no active service worker')
+                const newestSW = registration?.waiting || registration?.active
+                if (!newestSW) {
+                    resolve({ clientsCount: 0 })
+                    return
                 }
+
                 // Send request message to newest SW
-                const newestSW = registration.waiting || registration.active
                 newestSW.postMessage({ type: swMsgs.getClientsInfo })
                 // Resolve with payload received from SW `clientsInfo` message
                 this.offlineEvents.once(swMsgs.clientsInfo, resolve)
@@ -93,22 +194,43 @@ export class OfflineInterface {
      * or claim clients if it's the first SW activation
      */
     useNewSW() {
-        navigator.serviceWorker.getRegistration().then((registration) => {
-            if (!registration) {
-                throw new Error('No service worker is registered')
-            }
-            if (registration.waiting) {
-                // Update existing service worker
-                registration.waiting.postMessage({
-                    type: swMsgs.skipWaiting,
-                })
-            } else if (registration.active) {
-                // (First SW activation) Have SW take control of clients
-                registration.active.postMessage({
-                    type: swMsgs.claimClients,
-                })
-            }
-        })
+        if (!testSWAvailable()) {
+            return Promise.resolve()
+        }
+
+        return navigator.serviceWorker
+            .getRegistration()
+            .then((registration) => {
+                if (!registration) {
+                    throw new Error('No service worker is registered')
+                }
+                if (registration.waiting) {
+                    // Update existing service worker
+                    registration.waiting.postMessage({
+                        type: swMsgs.skipWaiting,
+                    })
+                } else if (registration.active) {
+                    // (First SW activation) Have SW take control of clients
+                    registration.active.postMessage({
+                        type: swMsgs.claimClients,
+                    })
+                }
+            })
+    }
+
+    /**
+     * @param {Object} params
+     * @param {Function} params.onUpdate - Called on status updates with argument { isConnected: bool }
+     * @returns {Function} - An unsubscribe function
+     */
+    subscribeToDhis2ConnectionStatus({ onUpdate }) {
+        if (!testSWAvailable()) {
+            return () => undefined
+        }
+
+        this.offlineEvents.on(swMsgs.dhis2ConnectionStatusUpdate, onUpdate)
+        return () =>
+            this.offlineEvents.off(swMsgs.dhis2ConnectionStatusUpdate, onUpdate)
     }
 
     /**
@@ -135,11 +257,10 @@ export class OfflineInterface {
         onCompleted,
         onError,
     }) {
-        if (!this.pwaEnabled) {
-            throw new Error(
-                'Offline features are not enabled. Make sure `pwa.enabled` is `true` in `d2.config.js`'
-            )
+        if (!testPWAAndSW()) {
+            return
         }
+
         if (!sectionId || !onStarted || !onCompleted || !onError) {
             throw new Error(
                 '[Offline interface] The options { sectionId, onStarted, onCompleted, onError } are required when calling startRecording()'
@@ -191,10 +312,8 @@ export class OfflineInterface {
      * @returns {Promise} A promise that resolves to an array of cached sections.
      */
     async getCachedSections() {
-        if (!this.pwaEnabled) {
-            throw new Error(
-                'Cannot get cached sections - PWA is not enabled in d2.config.js'
-            )
+        if (!testPWAAndSW()) {
+            return []
         }
 
         await navigator.serviceWorker.ready
@@ -219,10 +338,8 @@ export class OfflineInterface {
      * @returns {Promise} A promise that resolves to `true` if at least one of the cache or the idb entry are deleted or `false` if neither were found.
      */
     async removeSection(sectionId) {
-        if (!this.pwaEnabled) {
-            throw new Error(
-                'Cannot remove section - PWA is not enabled in d2.config.js'
-            )
+        if (!testPWAAndSW()) {
+            return false
         }
         if (!sectionId) {
             throw new Error('No section ID specified to delete')