+++ shell/src/App.js
@@ -7,19 +7,19 @@ const D2App = React.lazy(() =>
 ) // Automatic bundle splitting!
 
 const appConfig = {
-    url:
-        process.env.REACT_APP_DHIS2_BASE_URL ||
-        window.localStorage.DHIS2_BASE_URL,
+    url: process.env.REACT_APP_DHIS2_BASE_URL,
     appName: process.env.REACT_APP_DHIS2_APP_NAME || '',
+    appVersion: process.env.REACT_APP_DHIS2_APP_VERSION || '',
     apiVersion: parseInt(process.env.REACT_APP_DHIS2_API_VERSION),
     pwaEnabled: process.env.REACT_APP_DHIS2_APP_PWA_ENABLED === 'true',
+    plugin: process.env.REACT_APP_DHIS2_APP_PLUGIN === 'true',
 }
 
 const App = () => (
     <AppAdapter {...appConfig}>
         <React.Suspense
             fallback={
-                <Layer translucent level={layers.alert}>
+                <Layer level={layers.alert}>
                     <CenteredContent>
                         <CircularLoader />
                     </CenteredContent>